Function drm_ffi::mode::set_cursor2

pub fn set_cursor2(
    fd: BorrowedFd<'_>,
    crtc_id: u32,
    buf_id: u32,
    width: u32,
    height: u32,
    hot_x: i32,
    hot_y: i32,
) -> Result<drm_mode_cursor2>

👎Deprecated: use a cursor plane instead

Set cursor state (with hotspot position)

The buffer must be allocated using the buffer manager of the driver (GEM or TTM). It is not allowed to be a dumb buffer.

The hotspot position is used to coordinate the guest and host cursor location in case of virtualization.