drm_ffi

Module syncobj

Bindings for DRM sync objects

Functions

create

Creates a syncobj.

destroy

Destroys a syncobj.

fd_to_handle

Imports a file descriptor exported by handle_to_fd back into a process-local handle.

handle_to_fd

Exports a syncobj as an inter-process file descriptor or as a poll()-able sync file.

query

Queries for state of one or more timeline syncobjs.

reset

Resets (un-signals) one or more syncobjs.

signal

Signals one or more syncobjs.

timeline_signal

Signals one or more specific timeline syncobj points.

timeline_wait

Waits for one or more specific timeline syncobj points.

transfer

Transfers one timeline syncobj point to another.

wait

Waits for one or more syncobjs to become signalled.